pyproject.toml
[project]
name = "ragagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4",
    "langchain-core>=1.0.1",
    "langchain-huggingface>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.2",
]
